Partial refactor / reorganization into packages
