+++ app/build.gradle
@@ -2,7 +2,6 @@ apply plugin: 'com.android.application'
 
 android {
     compileSdkVersion 25
-    buildToolsVersion "25.0.1"
     defaultConfig {
         applicationId "com.hfad.bitsandpizzas"
         minSdkVersion 19
