+++ app/src/main/java/com/hfad/bitsandpizzas/ui/main/TopFragment.java
@@ -1,10 +1,12 @@
-package com.hfad.bitsandpizzas;
+package com.hfad.bitsandpizzas.ui.main;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import com.hfad.bitsandpizzas.R;
 
 
 /**
